+++ tests/test_run.py
@@ -1,8 +1,13 @@
+import logging
 from unittest import mock
 
+import pandas as pd
 import pytest
 
 import kotsu
+
+
+logger = logging.getLogger(__name__)
 
 
 class FakeRegistry:
@@ -25,10 +30,10 @@ class FakeRegistry:
         return self.entitys
 
 
-def test_form_results(mocker):
+def test_form_results(mocker, tmpdir):
     patched_run_validation_model = mocker.patch(
         "kotsu.run._run_validation_model",
-        side_effect=[({"test_result": "result"}, 10), ({"test_result": "result"}, 10)],
+        side_effect=[({"test_result": "result"}, 10), ({"test_result": "result_2"}, 20)],
     )
     patched_store_write = mocker.patch("kotsu.store.write")
 
@@ -37,27 +42,32 @@ def test_form_results(mocker):
     validations = ["validation_1"]
     validation_registry = FakeRegistry(validations)
 
-    kotsu.run.run(model_registry, validation_registry)
+    results_path = str(tmpdir) + "validation_results.csv"
+    kotsu.run.run(model_registry, validation_registry, results_path=results_path)
 
-    results = [
-        {
-            "test_result": "result",
-            "validation_id": "validation_1",
-            "model_id": "model_1",
-            "runtime_secs": 10,
-        },
-        {
-            "test_result": "result",
-            "validation_id": "validation_1",
-            "model_id": "model_2",
-            "runtime_secs": 10,
-        },
-    ]
+    results_df = pd.DataFrame(
+        [
+            {
+                "validation_id": "validation_1",
+                "model_id": "model_1",
+                "runtime_secs": 10,
+                "test_result": "result",
+            },
+            {
+                "validation_id": "validation_1",
+                "model_id": "model_2",
+                "runtime_secs": 20,
+                "test_result": "result_2",
+            },
+        ]
+    )
 
     assert patched_run_validation_model.call_count == 2
-    patched_store_write.assert_called_with(
-        results, "./validation_results.csv", to_front_cols=mock.ANY
+    pd.testing.assert_frame_equal(
+        patched_store_write.call_args[0][0],
+        results_df,
     )
+    assert patched_store_write.call_args[0][1] == results_path
 
 
 @pytest.mark.parametrize("artefacts_store_directory", [None, "test_dir/"])
@@ -98,3 +108,85 @@ def test_validation_calls(artefacts_store_directory, mocker):
                 ),
             ]
         )
+
+
+@pytest.mark.parametrize("skip_if_prior_result", [True, False])
+def test_skip_if_prior_result(skip_if_prior_result, mocker, tmpdir):
+    patched_run_validation_model = mocker.patch(
+        "kotsu.run._run_validation_model",
+        side_effect=[
+            ({"test_result": "result"}, 10),
+            ({"test_result": "result_2"}, 20),
+            ({"test_result": "result_3"}, 30),
+        ],
+    )
+
+    models = ["model_1"]
+    model_registry = FakeRegistry(models)
+    validations = ["validation_1"]
+    validation_registry = FakeRegistry(validations)
+
+    results_path = str(tmpdir) + "validation_results.csv"
+    kotsu.run.run(
+        model_registry,
+        validation_registry,
+        results_path=results_path,
+        skip_if_prior_result=skip_if_prior_result,
+    )
+
+    results_df = pd.DataFrame(
+        [
+            {
+                "validation_id": "validation_1",
+                "model_id": "model_1",
+                "runtime_secs": 10,
+                "test_result": "result",
+            },
+            {
+                "validation_id": "validation_1",
+                "model_id": "model_2",
+                "runtime_secs": 20,
+                "test_result": "result_2",
+            },
+        ]
+    )
+
+    out_df = pd.read_csv(results_path)
+
+    assert patched_run_validation_model.call_count == 1
+    pd.testing.assert_frame_equal(out_df, results_df.iloc[:1])
+
+    models = ["model_1", "model_2"]
+    model_registry = FakeRegistry(models)
+
+    kotsu.run.run(
+        model_registry,
+        validation_registry,
+        results_path=results_path,
+        skip_if_prior_result=skip_if_prior_result,
+    )
+
+    out_df = pd.read_csv(results_path)
+
+    if skip_if_prior_result:
+        assert patched_run_validation_model.call_count == 2
+        pd.testing.assert_frame_equal(out_df, results_df)
+    elif not skip_if_prior_result:
+        results_df = pd.DataFrame(
+            [
+                {
+                    "validation_id": "validation_1",
+                    "model_id": "model_1",
+                    "runtime_secs": 20,
+                    "test_result": "result_2",
+                },
+                {
+                    "validation_id": "validation_1",
+                    "model_id": "model_2",
+                    "runtime_secs": 30,
+                    "test_result": "result_3",
+                },
+            ]
+        )
+        assert patched_run_validation_model.call_count == 3
+        pd.testing.assert_frame_equal(out_df, results_df)