+++ tests/test_store.py
@@ -12,10 +12,12 @@ from kotsu import store
     ],
 )
 def test_write(to_front_cols, tmpdir):
-    results = [
-        {"id": "v1", "result_b": 2, "result_c": 3},
-        {"id": "v2", "result_b": 20, "result_c": 30},
-    ]
+    results = pd.DataFrame(
+        [
+            {"id": "v1", "result_b": 2, "result_c": 3},
+            {"id": "v2", "result_b": 20, "result_c": 30},
+        ]
+    )
     results_path = str(tmpdir) + "validation_results.csv"
     store.write(results, results_path, to_front_cols)